+++ onnx/__init__.py
@@ -3,7 +3,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .onnx_pb2 import *
+from .onnx_ml_pb2 import *
 
 import sys
