+++ onnx/numpy_helper.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 import sys
 
 import numpy as np
-from onnx.onnx_pb2 import TensorProto
+from onnx.onnx_ml_pb2 import TensorProto
 from onnx import mapping
 
 if sys.byteorder != 'little':
