+++ onnx/test/basic_test.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
-from onnx.onnx_pb2 import AttributeProto, NodeProto, GraphProto, ModelProto, IR_VERSION
+from onnx.onnx_ml_pb2 import AttributeProto, NodeProto, GraphProto, ModelProto, IR_VERSION
 import unittest
